Move xml2json to the right directory
